Deselect a previously selected day
A new feature is introduced in the component that allows to deselect a previously selected day. Before this improvement, if a user clicked again on the same selected day, no change occurred. With this change, it is now possible to deselect the selected day by setting the model value to null.